コマンドファイルを分割
