+++ src/components/milkdown/commands/index.ts
@@ -1,0 +1,6 @@
+export * from './moveDown';
+export * from './moveLeft';
+export * from './moveRight';
+export * from './moveUp';
+export * from './selectLineEnd';
+export * from './selectLineStart';
