+++ src/components/milkdown/commands/selectLineEnd.ts
@@ -1,0 +1,4 @@
+import { selectTextblockEnd } from '@milkdown/prose/commands';
+import { $command } from '@milkdown/utils';
+
+export const selectLineEnd = $command('select-line-end', () => () => selectTextblockEnd);
