+++ src/components/milkdown/commands/selectLineStart.ts
@@ -1,0 +1,4 @@
+import { selectTextblockStart } from '@milkdown/prose/commands';
+import { $command } from '@milkdown/utils';
+
+export const selectLineStart = $command('select-line-start', () => () => selectTextblockStart);
